fix(merchant acquirer): change contract db

+++ services/merchant-acquirer/src/main/resources/db/migration/V1__create_merchants_table.sql
@@ -4,6 +4,6 @@ create table Merchants (
     mcc CHAR(4) NOT NULL,
     category VARCHAR(50) NOT NULL,
     acquirer_id VARCHAR(50) NOT NULL,
-    acquiring_fee NUMERIC(5, 4) NOT NULL,
-    average_check BIGINT NOT NULL
+    acquiring_fee NUMERIC(5, 0) NOT NULL,
+    average_check NUMERIC(5, 4) NOT NULL
 );
